refactor : Money.java
 - 수익 double 로 리팩터링

+++ src/main/java/domain/participant/Money.java
@@ -11,21 +11,21 @@ public class Money {
 		return new Money(money);
 	}
 
-	private int validateMoney(String money) {
-		int moneyAmount = getValidatedMoney(money);
+	private double validateMoney(String money) {
+		double moneyAmount = getValidatedMoney(money);
 		validateAmount(moneyAmount);
 		return moneyAmount;
 	}
 
-	private void validateAmount(int money) {
+	private void validateAmount(Double money) {
 		if (money < 0) {
 			throw new IllegalArgumentException("베팅 금액은 음수가 아니어야 합니다.");
 		}
 	}
 
-	private int getValidatedMoney(String money) {
+	private double getValidatedMoney(String money) {
 		try {
-			return Integer.parseInt(money);
+			return Double.parseDouble(money);
 		} catch (NumberFormatException e) {
 			throw new IllegalArgumentException("올바른 형식의 입력이 아닙니다.");
 		}
